+++ packages/lib/src/lib.ts
@@ -1,5 +1,5 @@
 export * from "./types";
-export * from "./helpers/efficient_ecdsa";
+export * from "./helpers/public_input";
 export * from "./core/membership_prover";
 export * from "./core/membership_verifier";
 export * from "./helpers/tree";
